pyproject.toml
[project]
name = "python-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py[voice]>=2.6.4",
    "python-dotenv>=1.2.1",
    "spotipy>=2.25.2",
    "yt-dlp>=2025.12.8",
]
